SellerApi/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
boto3==1.34.34
python-dotenv==1.0.0
email-validator==2.1.0
protobuf==4.25.2
